Update workarounds for potential javac type inference issue.
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@980755 13f79535-47bb-0310-9956-ffa450edef68

+++ tutorials/src/org/apache/pivot/tutorials/webqueries/ExpensesWindow.java
@@ -238,7 +238,7 @@ public class ExpensesWindow extends Window implements Bindable {
     @SuppressWarnings("unchecked")
     private void updateSelectedExpense() {
         Object expense = expenseTableView.getSelectedRow();
-        final Integer id = JSON.get(expense, "id");
+        final int id = (Integer)JSON.get(expense, "id");
 
         expenseSheet.load(expense);
         expenseSheet.open(this, new SheetCloseListener() {
@@ -291,7 +291,7 @@ public class ExpensesWindow extends Window implements Bindable {
     @SuppressWarnings("unchecked")
     private void deleteSelectedExpense() {
         Object expense = expenseTableView.getSelectedRow();
-        final Integer id = JSON.get(expense, "id");
+        final int id = (Integer)JSON.get(expense, "id");
 
         deleteConfirmationPrompt.open(this, new SheetCloseListener() {
             @Override
